PNEUMATICS and HYBRID WHEELS Update
PneumaticsAPI has been renamed "Pneumatics" and consists only of
compressor controls. This change means that the API for each component
will have to consist of solenoid controls for the solenoids that they
are using. This change is reflected in the updates to the Hybrid Wheels
API, which now has controls for its solenoids in addition to its wheel
controls.