Autonomous Updates and Gyroscope Beginnings
This commit's main purpose is to update the autonomous programs with the
capability to dynamically change given changes in switch and scale color
randomization as well as the position of our robot. Additionally, it
includes the code for the gyroscope, but the gyroscope is not currently
functional given its orientation on the robot.